pyproject.toml
[tool.black]
line-length = 88
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
)/
'''

[tool.poetry]
name = "graphene-django-extras"
version = "0.5.2"
description = "This library add some extra funcionalities to graphene-django to facilitate the graphql use without Relay, allow paginations and filtering integration and add some extra directives"
readme = "README.md"
authors = [
    "Ernesto Perez Amigo <eamigop86@gmail.com>",
    "Martin Andersen <dk.mi.andersen@gmail.com>",
    "Lukasz Dynowski <ludyn@leo-pharma.com>"
]
license = "MIT"
homepage="https://github.com/eamigo86/graphene-django-extras"
repository="https://github.com/eamigo86/graphene-django-extras"
documentation="https://github.com/eamigo86/graphene-django-extras"
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: PyPy",
]
keywords=["api", "graphql", "protocol", "graphene", "django"]

[tool.poetry.dependencies]
python = "^3.6 || ^3.7 || ^3.8 || ^3.9"
django-filter = "^21.1"
djangorestframework = "^no"
python-dateutil = "^2.8.0"
graphene-django = {version = "^3.0.0-beta.7", allow-prereleases = true}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^18.3-alpha.0"
tox = "^3.7"
factory_boy = "^2.11"
pytest-django = "^3.4"
ipdb = "^0.13.9"
flake8 = "^3.6"
bandit = "^1.5"
pytest-cov = "^2.6"
pyflakes = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
